feat(infra): iam, node and cluster helpers for k8s module
Module function creates to create aws iam roles needed for eks cluster.
The admin role requires mfa to assume it will be mapped to the k8s
cluster-admin. The automation role is only assumable by ec2 instances
and will be mapped to an k8s automation role. Lastly, the nodeGroupRole
will be assigned to the managed node groups and have the AWS managed
policies required by EKS.

Module function that creates a managed node group in a specified subnet.
The subnet must be the private subnet for our VPC.

Module function to create the EKS cluster. This creates a KMS key which
the cluster will use for envelope encryption. An cluster OIDC provider
is created which will be needed by cluster autoscaler. Lastly, the admin
aws iam role is mapped to the k8s system-masters group. This means the
admin role has full access to k8s resources. The automation role is
mapped to the automation group. The cluster role binding for this group
will come in a later commit.