feat(infra): provision eks cluster with one managed node group
An eks cluster is provisioned in eu-west-1. The control plane is
deployed in an AWS managed VPC. We then deploy 1 managed node group in
our own VPC.

The end state we want our EKS cluster in is to have the cluster endpoint
only accessible from within our VPC. However, pulumi needs access to the
cluster to configure the `aws-auth` configmap and VPC CNI add on.
We therefore need to first provision the cluster with public access to
the cluster endpoint. This will allow us to configure the cluster.
Once this is done we then disable the public access.