feat(infra): kubernetes module functions to secure iam
This introduces more functions that help us conform to EKS best
practices for identity and access management.

* Function that updates the aws-node service account and
daemonset to use IRSA (IAM Roles for Service Accounts). An AWS role that
uses the managed EKS_CNI policy is created for this.

* Function that creates a restricted pod security policy and applies it
the service accounts in the kube-system namespace as well as the
automation group.

* Function to create the k8s automation. It is allowed access to k8s
resources for deployment purposes.

( Helper function to create an automation role binding for a
particular namespace.