feat(cockroach): initialize service
initialize the usage of cockroachdb in our services' folder.
This allows cockroach to be used against our dev cluster initially.

Kustomize is being used for applying the kubernetes config to the cluster. It allows easy use of overlays for different environments. Initially only dev is used.

The .gitignore is added to not commit temp certs created for cockroachdb