feat(cockroach): add kratos db and change setup
Changed the way cockroachdb admin user and setup is done. The previous system deployed a new client in the cluster. Moved to a port-forward and running an init script instead. In the future, this can be moved to pulumi provider.

Added the setup of the kratos user and database with permission

Signed-off-by: Matt de Haast <matt@fynbos.dev>