feat(protea): get user session when app loads
Every page is SSR where we get the session info from Kratos. We call out
to the `whoami` endpoint on Kratos to check if the session
is still valid. The session object returned from Kratos is passed to the
app props. Session expiry is not handled yet.

An auth provider is introduced so that the current session can be
retrieved with the `useSession` hook in our components.