feat(infra): kratos development config
Kratos identity schema that defines a user that uses email and
password verification.

Kratos config that is set up to use localhost:3000 as the ui.
Signup, login, verification, recovery and settings flows are
enabled.

Docker-compose config for local dev that spins up kratos, runs
its migrations and a smtp server so that verification and
recovery emails can be sent and seen locally.