feat(protea): move session out of _app
This removes the auth context provider that was providing the session
information to pages. And rather fetches the session data in
getServerSideProps on each page.

To achieve SSR, we need a second kratos client to be used server side.
This is only because we can't resolve dns inside the container - the
client itself does actually work both on the client and server.

There are two default functions exposed for getServerSideProps:

- checkSession: to be used on pages that shouldn't need auth, and should
                be redirected away from when the user has an auth context.
                e.g. login
- getSession:   to be used on pages that should be authed, redirects to
                login/signup if the user isn't authed.