feat(backend): start graphql endpoints for organisation
The service has been wired up to connect to cockroachdb so that the e2e
flow can be tested for creating and getting organisations using graphql.
The e2e flow spins up CockroachDb in a test container and serves the
graphql endpoint using the go test http server. An actual graphql client
is then used to make requests against the server and create/read
organisations from the DB.

Note that when connecting to the DB we read in the certs from file and
build the connections string with the ssl inlined to side step file
permission issues when it comes to deploy time.