feat(backend): organisations service and testing utils
The testing workflow is to use test containers to
start a single-node CockroachDB instance with ssl disabled.
The backend database is created and the migrations applied.
We run the tests in serial for now and truncate the db after
each test. Since the tests are run in serial we use a global
db connection.

This workflow has been applied to the unit tests for the
organisation service.