feat(backend): add authorization to services
We use Oso to define authorization policies which the organisation
service will enforce. In this case, an org can only be read by the
its owner.

The user service is introduced to get a user from the kratos cookie.
This integration is tested using kratos in the
docker-compose environment. A mock user service exists for tests.