feat(ci): use buildkit to build images and cache from ecr
Kaniko was tested to do the builds and push to ECR. However, it was
found to be slow for multi-stage builds even with remote caching. Since
the backend base image needs to be run for testing anyway, a custom
docker image is built with docker-ecr-credential-helper baked in (
new pulumi `ecr` project). This image is then used to build the images
with BUILDKIT. Although the docker layer caching isn't working properly
with BUILDKIT and ECR, it still has the same run time as Kaniko with
caching.

The custom image is used to avoid running custom userdata on the runner
instance to install docker-ecr-credential-helper and mount it into
all containers running on the instance. Instead, the manager only has
docker-ecr-credential-helper installed allowing it to pull the image
only when it's specified in the `.gitlab.yaml` files.

refs: https://gist.github.com/dreampuf/a0d416a15299a2ac74a0a5cb8f2871c0