feat(protea): preview apollo middleware
Allows setting the fynbos-preveiw header on all apollo queries based on
the preview status. This can be consumed by the control plane.

One caveat is that the preview status needs to be passed to the apollo
client as context. This is wanted as we don't always want/need the
preview header.