feat(ci): set up deploy pipelines
The trigger pipeline has 5 stages: build_base, test_base, publish,
deploy_preview and deploy_live. For each of these stages, it will
trigger the corresponding stage in each service's child pipeline.

The build_base stage builds an image that has the dependencies installed
and pushes that to the appropriate ECR. The test_base stage will then
pull the base image and test it. Publish builds the production image
for use in the deploy stages.

The deploy stages will assume the `eksAutomationRole` which will allow
it to pull the kubeconfig for the relevant environment. Currently, a
`kubectl version` is run to make sure it can connect to the eks cluster.
This will be substituted with the appropriate deploy commands.