feat(pacioli): create accounts in TigerBeetle
Tenants can create accounts in ledgers. The accounts inherit its `code`
from an account category. This will be used later on when transfers
are created to enforce transaction types. The `unit` needs to be
assigned which represents the currency.

UUIDs are used for account IDs and the string format is converted back
into a `u128` so that it can be stored in TigerBeetle. The `ledgerID` is
stored in the `userData` field of TigerBeetle so that it can be
retrieved later to enforce tenant isolation.