refator(pacioli): create and get transfers
Transfers are created and retrieved in batches. For now it is assumed
that the transfer IDs are uuids and generated outside the service. It
might allow for more use cases if we are un-opinionated about the ID
format and just accept `[]byte`. Since we are using batching, the grpc
api only returns the errors in the batch.

The service inherits the error set of TB as it is well defined.