feat(infra): dev k8s cluster
update the k8s dev cluster to get it to a working state which includes add CSI, fixing IAM permissions, updating the cluster size to 2 nodes, removing calico and adding the automation role to k8s to allow for deployments.

add the deployment of EBS CSI plugin. This allows pods to be able to provision persistent volumes with EBS that are encrypted by default

Calico was removed as it was causing network errors that were not easily resolved.

Signed-off-by: Matt de Haast <matt@fynbos.dev>