feat(protea): create wallet layout component
- Creates the FAB
- Allows custom classes on the Router's anchor, this was needed to allow
the NavBar buttons to function as expected.

The WalletLayout component forms a scaffold, to be used as a root
component on all pages within the wallet (Doesn't include login/signup/
verify/marketing pages). The component includes the NavBar, NavRail and
NavDrawer components (And allows hiding them). It also creates a new
version of the header component.
The children of the WalletLayout will be wrapped in a grid which matches
the breakpoints in the designs.