feat(infra): kratos dev cluster
fix kratos to be able to work in the dev cluster by dynamically
applying the domain on which it is deployed. The deployment for local
and dev cluster has been updated as well.

The ability to set the GIT_HASH for dev cluster has been added to
allow you to update the cluster using a set hash for the deployments

Signed-off-by: Matt de Haast <matt@fynbos.dev>