feat(infra): ci docker and ami update
upgrade the CI process to use docker 20.10. This includes the docker
version running on the runners and those in the base docker image.
The base image was also upgraded to ubuntu 21.10 to support the need
for a newer kernel for testing tigerbeetle in CI. The base AMI image
for our runners was also updated to ubuntu 21.10

Signed-off-by: Matt de Haast <matt@fynbos.dev>