fix(pacioli): fix testing
fixing the testing setup for pacioli by removing the requirement of
using bind mounts as this does not play well in our CI where we
mount the docker socket to do docker in docker. This has been replaced
by rather inline doing the init and start commands.

The connection string was also updated to use the `hostIP` to ensure
the testing for docker in docker works.

Signed-off-by: Matt de Haast <matt@fynbos.dev>