feat(backend): identity stores the kyc data for providers
The kyc data we collect will support Dwolla and will take minor
additions to support Nium. We are storing it on the `identity` table
for now for simplicity. We only need a subset of the kyc data to create
an identity. The rest must be supplied
when the identity is verified.

The identity service has a private `identity` model to store/get db rows
and exposes an `Identity` model to consuming services. This way we can
translate to the required `pq` driver types opaquely.