refactor(protea): kratos to work with remix
To align with how remix works, the kratos flows had to be refactored to
use the SSR browser flows. This required the hooks that were used to
initialise/get the flows to be replaced, so that they could be run in
the remix data loader. The form submission mechanisms were replaced
with remix actions.