feat(protea): reimplement the Select component
Implement the Select headless UI component.
https://headlessui.dev/react/listbox

NOTE: In order for the Selects value to be submitted by the remix Form,
a hidden input must be placed in the Form alongside the Select. The value
for the Select should be stored as a state variable to ensure reactivity.
The hidden input was not implemented inside the Select so as not to
prescribe what part of the Select's data should be submitted.