feat(protea): login challenge flow
A user will now be directed to /login/challenge if they are trying to
change settings that require a privileged session. Uses user_settings
cookie to tell the challenge flow where the user is coming from, and to
set a snackbar notification once the user has successfully set a new password.
Introduces a separate page for settings a new password after account recovery.