feat(backend): admin user service
Admin user service used to auth admin grpc endpoints. For production,
this expects the IdToken issued by the Google oauth2 service to be
present in the GRPC metadata. It then verifies the token and extracts
the email address otherwise it will return an error.

For tests, a mock user service is used that returns the string set in
the grpc metadata. Tests thus require the use if `ActingAs` helper
function to set the metadata.