feat(protea): setup flow control
Flows will allow users to perform a specific task that requires multiple
steps. Currently this flow is stored in the user-settings cookie, but
should be moved to a db table at some stage. The user-settings cookie
max age has been increased to a day to accommodate this.

The following mechanisms are provided to manage a flow:

- Initialization of flows
  - Ensures there is no flow with the current ID.
  - Will check for existing flows of the same type.
  - Ensures that the array of flows can be treated as a stack by the
    other methods that access the flow.
- Fetching the current flow
- Stepping through the flow (saving data on the way)
- Updating the data of a specific flow step
- Completing a flow
  - Marks the current flow as complete.
- Exiting a flow
  - Removes the current flow from the stack and returns to the previous flow
    (Or exits if there is no previous flow).