feat(protea): infinite scroll on activity page
Had to make use of a remix action to route to an individual transaction
page, so that the scroll restoration would work. This allows us to set a
cookie which tells the activity page how many pages were loaded previously.

The infinite scroll works by watching the scroll position, and calling a
load of the loader with a fetcher.
https://remix.run/docs/en/v1/api/remix#usefetcher

Some of the useEffects don't have exhaustive deps (and now have linting
disabled for those lines), because this would cause unwanted side effects.