feat(infra): deploy peer rafiki to local cluster.
Deploy a peer rafiki instance to the local cluster so that outgoing
payments can be tested.

Update Postgres deployment to the "postgres" super user to run an
init script that creates databases and users for rafiki and peer
deployments.

The peer and rafiki deployments use separate redis databases on the same
instance.