infra(dev): bottlerocket nodegroup
the previous node group became unmanageable as the launch
template was manually editted to allow kernel 5.10 to be enabled.
This meant that we could no longer manage the node group through
aws/pulumi as it kept failing. This commit updates the pulumi
version to support a managed node group of bottlerocket. The
new nodegroup is created with bigger instances as well.

Signed-off-by: Matt de Haast <matt@fynbos.dev>