feat(protea): add initiate onboarding to verify
This ensures that the unit application form is created during the
verification step. Currently you will need to route back to /verify after
successfully verifying your email, to route to the application form, as
we have no way to detect the unit account yet. This will be fixed with
the flow updates.