feat(backend): handle validation for the grpc api
Adds validation to the grpc resolver code, which will surface validation
errors cleanly to the client. If validation errors are thrown inside the
underlying services, they will be surfaced to the grpc client as internal
errors.

This provides separation of concerns such that we don't overload the
service validation. The service validation is now used to ensure data
integrity.

Adds validation to the UpdateOnboarding rpc resolver.