refactor(backend): unit owns creating and getting of unit counter party
Move the creating and getting of the counter party into the unit provider
and out of the funding source service.
Our counter party model maps the unit id to our funding source. At the
moment we assume a funding source can only have 1 counter party and
is enforced with a composite primary key in the database.