feat(infra): dev cluster cockroach certs
setup the dev clusters cockroachdb node certs.
This requires setting up a role on the pki cert
called `crdb-node` that allows signing for the
various allowed domains that crdb node requries.

Then a new role `crdb-node` is added to the kubernetes
auth for `dev-euw1` that gives permission for the role
to sign pki certs `pki/dev-int/sign/crd-node` for the
pki role. This role scopes it to only be allowed in the
`cockroachdb` namespace.

Finally, the cockroachdb namespace, service account and
issuer are created that references the role in vault to
create certs.