refactor(backend): configure the unit on boarding workflow to use signals
- Removes country dependency when creating the workflow (We are already
  calling out to the identity service where Country was require).
- Creates a new activity for creating a unit application.
- Renames the workflow args to state. The args passed into a temporal
  workflow are persisted for the lifetime of the workflow. Calling them
  state makes this more obvious.

The workflow is now configured to create a unit application.
This application can be immediately approved or denied. If the
application is not immediately approved or denied, the workflow will
wait indefinitely for webhooks that approve or deny the application.

If the application is approved (immediately or through webhooks), a
fynbos account is created and mapped to the unit customer.

An account is not created if the application is denied.