feat(backend): mx service can get account balance
The mx service will start the balance aggregation. It returns the member
so the calling service will have to decide whether this was successful
or not.
After the aggregation is complete the mx service can get the balance. It
will convert this to a struct that details the assetCode, assetScale
and balance (as a int64).