RFQ for structure of project
Pacioli
Deposits
Accounts

Have been updated. It's kind of hard to see it now, but basically all constructors will just take in Backends as an interface with all dependencies. Each individual service will take a subset of the main Backends and type inferance will allow it to use only the subset it needs.

This also doesn't compile but I'd rather get feedback early and ongoing as I complete the migration