feat(backend): insert agreements in init container
As @barnarddt suggested, inserting
agreements everytime a container
starts up as part of the agreements
service creation is not a great idea.

I had two options to overcome this
problem:

1) Making the function idempotent
so it won't be trying to insert same
agreements over and over again.

2) Running the function as part of
an init container.

Second option made more sense to me
since agreements won't be updated
frequently.