feat(ci): runner s3 cache
setup the s3 runner cache for the k8s runner. This
required setting up the role and service account
manually and binding the policies to that. We are
currently using the s3 bucket that is from the older
runner. We need to migrate the off at a later stage.

Also added some compression variables for the pipeline
that speeds up compression.