refactor(backend): createMxAccount workflow only creates the mx account
The creating of the fundingsource and unit counterparty has been removed
from this workflow. It is now only concerned about creating the mx
account.

The `GetBankDetails` grpc endpoint now waits for the workflow to have
completed before looking up the bank account.

The above better supports the designs where the bank details are shown
to the user before the user provides an OTP to confirm the linking.