feat(protea): better layout handling
Makes use of the remix handle to define what app layout to use on a per
route basis. This is better than having a global route map object. The
remix handle can export a function, which we are using to allow showing
different routes if the user is signed in or not (only currently used
at /).

https://remix.run/docs/en/v1/api/conventions#handle

We will use a similar mechanism to fix the back button routing.

Also cleans up a lot of the "coming soon" pages. And formats the
verification page.