feat(protea): allow programmatic routing with back button
By default the back button on focused flows will pop the history stack.
If `backTo` is defined in the loader of a page, then the back button
will route there rather. This is useful for the end of flows that end up
at pages that can be routed to from elsewhere.