feat(protea): linked-account almost there step
This adds an almost there step during the adding of a linked account.
Which allows the flow to wait until the linked account has actually been
added.

We also animate some of the shapes in the flow.

This also improves the exit and entry to the flow based on where the
user comes from. Clicking the topup button without a linked account will
now redirect to topup afterwards.