refactor(backend): token based pagination
Replaces the current page number based pagination with a token based
version. This roughly follows the google spec:
https://cloud.google.com/apis/design/design_patterns#list_pagination

This does not implement opaque tokens, and just uses the ID of the
underlying data types ID for the token.

Removes a bunch of unnecessary endpoints instead of upgrading them, as
they were redundant or deprecated.