refactor(protea): wallet layout header
Removes the fynbos logo from the header on each page, and adds the page
title to the header rather than on the first card.

The fynbos logo is still used in places where there is no page title.
e.g. the home page.

This also refactors the HomeShapes as these are now simpler.