feat(protea): extend current error handling method for connect client
Creates our own version of the ConnectError with fields we require to
be able to provide the error handling we need.

This now has field validations built in to the error instance and will
be able to handle PreconditionFailure violations in the future.

There is now a helper on the error response for throwing http errors
and includes its own version of the error function (For throwing
snackbars etc.).

Refactors the sentry captureMessage to be called when the connect error
is instantiated.